requirements.txt
openenv-core==0.2.3
fastapi>=0.111.0
uvicorn>=0.29.0
pydantic>=2.0.0
openai>=1.30.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
